Refactor Organization model into separate app
