+++ members/admin.py
@@ -1,5 +1,4 @@
 from django.contrib import admin
-from .models import Member, Organization
+from .models import Member
 
-admin.site.register(Organization)
 admin.site.register(Member)
